fixup! split A.3, abci-in-rpc
