+++ rpc/src/serializers/hash_base64.rs
